+++ cli/docs/changelog.md
@@ -1,5 +1,10 @@
 # TBD
 
+# 0.5.1
+### Features
+* Add instructions for users on what to do if no Kurtosis engine is running
+* If an engine isn't running, the CLI will try to start one automatically
+
 # 0.5.0
 ### Changes
 * Replaced `EnclaveManager` with `Kurtosis Engine API Libs` which handle all the interactions with the `Kurtosis Engine Server`